fix: Show DSN connection string format first (#19460)
#### Summary

The issue of users unable to connect to Postgres because they need to encode their password value in the URL when it contains special characters has come up multiple times before. An easy way to avoid it is to use the DSN format. It's much easier I think to quote a value in the DSN format (to handle spaces) vs percent encode special URI characters.

I changed the docs a bit to list the DSN format first.

Feel free to close if this doesn't make sense to you