fix: Renames tables before writing summary. (#19600)
The Basic Transformer introduced the concept of renaming tables.

This works well, but an unfortunate side-effect of it is that sync_v3 had to be aware of the renames, because it does post-processing things to the tables, like delete-stale.

Because of this, a `tableNameChanger` was implemented back then, which is used for all that happens after the sync.

At the time this was implemented, we didn't use to write table names to the sync summary.

When we added this feature, we forgot to replace table names. This PR does just that.

Here's a before/after sync summary row:
<img width="513" alt="Screenshot 2024-11-14 at 16 33 31" src="https://github.com/user-attachments/assets/f3158cbb-c851-423b-a699-69b646ba2f6d">