feat: Implement custom ORDER BY clause support. (#19674)
This PR implements support for setting custom ORDER BY clauses on any table or group of tables in the ClickHouse destination.

Because tables are defined on the source side and every destination can have different optimal indexing/sorting circumstances, it is not practical to define destination-specific sorting strategies at the source. Therefore, this enables setting it at the destination.

<img width="532" alt="Screenshot 2024-11-19 at 16 08 06" src="https://github.com/user-attachments/assets/4f447521-504d-42ff-a51e-5957d11da9ea">
<img width="445" alt="Screenshot 2024-11-19 at 16 08 32" src="https://github.com/user-attachments/assets/1b24c17c-427e-4915-9b3f-1d10c51512fa">