fix: Add timestamp parsing logic to JSON flattener (#19686)
fixes https://github.com/cloudquery/cloudquery-issues/issues/2821

The timestamp field parsing logic on the JSON Flattener is just plain wrong: it was assuming that timestamp type subfields would be of type `time.Time`, and this is not the case.

The `TypeSchema` reports timestamp types as `timestamp[us, tz=UTC]`, but in practice I've already seen two slightly different date formats.

Therefore, in this change, for timestamp subfields, if the underlying type of the subfield is `string`, we will try to parse the string datetime using a variety of popular formats, including some reported by our customers and some seen in the AWS plugin in tests.