feat: Add "fail immediately" spec option to test source plugin (#19863)
Adds a `fail_immediately` spec option to the test source plugin.

Initially I also added a `fail_after_n_resources` option, but it's
tricky to implement with multiple tables, and it would interact weirdly
with batching. I didn't see much benefit to be honest.

This will be used to add end to end testing at the CLI stage.