feat: Support all batch writers in test destination plugin. (#19964)
Adds spec options for all 3 batch writers on the test destination plugin.

My tests show that all 3 can be used. However, the streamBatchWriter CLI tests are getting stuck. This shouldn't block this PR, on the contrary, it might be good to add tests to the CLI using the `streamBatchWriter` and figure out why it's getting stuck.

Testing with this spec
<img width="598" alt="Screenshot 2024-12-13 at 16 55 26" src="https://github.com/user-attachments/assets/4659be2c-4949-4542-9b1f-387729da1eaf" />

Test pass when omitting the streamBatchWriter
<img width="1414" alt="Screenshot 2024-12-13 at 17 34 44" src="https://github.com/user-attachments/assets/94e17aa0-2ffe-4f62-bdc0-e493e8a36386" />