feat: Add rename_column to basic transformer plugin (#20025)
#### Summary



Related to https://community.cloudquery.io/t/basic-transformer-plugin-rename-column/985

This PR adds a `rename_column` feature to the basic transformer.
My use case is that I want to rename the `id` from the source table to `external_id` as I want to keep my own ids in the destination tables.

Note: I ran `make gen` and it added `| github.com/apache/arrow/go/v13 | Apache-2.0 |` line to the license file.
I did not add anything related to this, should I commit it still? It doesn't seem to be included in the plugin