feat: Publish UI Bundle Archive (#20116)
This PR creates and uploads a `bundle.tar.gz` of UI assets as part of the plugin publishing process. 

The bundle file will be downloaded and extracted in cases where we want to download/mirror all the files needed to serve a UI for a plugin. Otherwise, we don't have a way to list all the files.