feat: Add skip_incremental_tables suboption for partition_by. (#20276)
This PR adds a `skip_incremental_tables` boolean sub-option to the
`partition_by` strategies of the ClickHouse destination.

The reasoning is that it is a common use case that
incremental/non-incremental changes have to be partitioned differently,
but since tables can migrate from being non-incremental to becoming
incremental, the "list of incremental tables" can change over time, and
so maintaining this list in the spec up to date can be an error-prone
operational burden.

By adding the `skip_incremental_tables` options on a partitioning
strategy, a partitioning strategy can be applied specifically to
non-incremental tables, without needing to list the tables.