feat: Support migrating addition of `_cq_client_id` when order by is set (#20544)
#### Summary

When someone enables (for supported plugins) the `add_cq_client_id: true` spec option migration will fail at the moment as we're adding a new non null column.

This PR supports such migration with the caveat that it only works if `order_by` was already configured explicitly, otherwise `_cq_client_id` will be implicitly considered in the new sort key as it's a non null column and we can't do sort key changes (yet). Technically we can modify the sort key when adding a new column, but we'd have to make sure it's a prefix of the primary key https://clickhouse.com/docs/sql-reference/statements/alter/order-by so don't think it's worth supporting that.

The operation is fast to do, as it changes metadata only, and we don't need to set a default value. It will use the zero value of the string type, see https://clickhouse.com/docs/sql-reference/statements/alter/column#add-column. The data will be filled upon reading from the table as needed.