feat: Enable Role Assumption for S3 Destination (#20916)
#### Summary

Prior to this change a user could not assume a role without first generating a config file and using `local_profile`.  This change enables all of the role assumption that is possible in the AWS Source plugin, without the need for a local config file.

This is fully backwards compatible