feat: Support the new OTEL metric format removing the client_id metric point granularity (#21048)
Should go with https://github.com/cloudquery/plugin-sdk/pull/2185 in order to allow the CLI to read metrics without the client_id metric point.

[old CLI, old SDK](https://github.com/user-attachments/files/21236264/before.txt)
[old CLI, new SDK](https://github.com/user-attachments/files/21289151/after.txt)
[updated CLI, old SDK](https://github.com/user-attachments/files/21289612/after.txt)
[updated CLI, new SDK](https://github.com/user-attachments/files/21236263/after.txt)

The above showcase that Duration metrics are unavailable if using an old CLI version with newer plugin versions, but that's a downside we have to work with.