fix: Fix bug in ClickHouse destination where tables always get recreated (#21099)
There was a bug in the ClickHouse destination that caused this log line to always show, even if the table already existed:

```
Table doesn't exist, creating
```

Generally this didn't have much of an impact, because the `CREATE TABLE` statement called uses `CREATE TABLE IF NOT EXISTS`, so it was a no-op. However, the log line is incorrect, and the flood of unnecessary DDLs could have an impact on large clusters. This change fixes the bug.