feat: Add support for table-level TTLs in ClickHouse destination (#21101)
This adds support for table-level TTLs in the ClickHouse destination. 

Similar to partitions and ORDER BY clauses, a TTL section can be added to the config like this, so every table's TTL can be specified individually, if desired:

```
ttl:
- tables: ["*"]
  skip_tables: ["no_ttl_table"]
  ttl: "INTERVAL 60 DAY"
```

Some aspects of this were harder to do than anticipated. For example, ClickHouse transforms the TTL expression to a different format when it is stored in the database, so to avoid unnecessary DDLs we do a comparison of normalized expressions inside ClickHouse to evaluate if two TTL expressions are, in fact, equal.