test: Add test to validate progress reporting to API (#21116)
<!-- 🎉 Thank you for making CloudQuery awesome by submitting a PR 🎉 -->

#### Summary

Follow up to https://github.com/cloudquery/cloudquery/pull/21108.

This was a bit tricky as we need to list and download plugins from the
real API, and mock the progress reporting endpoint as we don't really
have a sync run to report progress for.

<!--
Use the following steps to ensure your PR is ready to be reviewed

- [ ] Read the [contribution
guidelines](https://github.com/cloudquery/cloudquery/blob/main/CONTRIBUTING.md)
🧑‍🎓
- [ ] Run `make lint` to ensure the proposed changes follow the coding
style 🚨 (install golangci-lint
[here](https://golangci-lint.run/usage/install/#local-installation))
- [ ] Run `make test` to ensure the proposed changes pass the tests 🧪
- [ ] If changing a source plugin run `make gen` to ensure docs are up
to date 📝
- [ ] Ensure the status checks below are successful ✅
--->

---------

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>