fix: Provide visual feedback of chat completion. (#21386)
The delay in responses on AI Onboarding can be confusing. This is
because there's a lot of back and forth between CLI, Cloud API and
OpenAI API. Also, for some Cloud API operations, there's many requests
due to pagination. So the user may experience delay for some operations,
and no feedback can be confusing.

This PR implements visual feedback as seen on this video:


https://github.com/user-attachments/assets/1338f112-53e9-4d13-a540-6d1cfc839a3c

---------

Co-authored-by: Kemal <223029+disq@users.noreply.github.com>