fix: Table with PK components should upsert. (#21474)
`schema.Column` defines both `PrimaryKey` and `PrimaryKeyComponent` for a column, and certain plugin's tables use one of the two but not the other.

The Postgres destination decides whether to insert or upsert based on the existence of PK columns on the table, but ignores PK components.

This causes tables with PK component only columns to fail with constraint violations on inserts, when the rows already exist on the DB side.