feat: Add create-sql-file tool. (#21494)
Add the `create-sql-file` tool to the AI Onboarding toolkit.

This tool is identical to the existing `create-spec-file`, with the only difference being semantics and the fact that the extension is `.sql` instead of `.yaml`.

There's already a merged function on cloud that provides example queries.

The last step after this is to change the system prompts so that the LLM can start leveraging them.